Ban Lombok.
Even the authors of Lombok admit it is controversial.  They willfully
violate the restrictions on how annotation processors are designed to be
used, and code using Lombok is no longer valid Java.  It cannot be
compiled without the annotation processor, so marking the plugin as
"optional" is incorrect and misleading.

Lombok is designed to hide boilerplate code behind opaque code
generation.  But one of the primary reasons that Java has been so
successful in the enterprise is precisely that it is explicit about what
it is doing.  Hiding generated code is antithetical to this.  A
developer should be able to easily confirm precisely how the final code
behaves, whether the equals method allows inherited classes or not, what
the format of the generated toString will be, etc.

Everything that Lombok provides is already provided by all modern IDEs.
Getters, setters, equals, hashcode, toString, and constructors are all
generated by Eclipse, IntelliJ, NetBeans, etc., with only a few
keystrokes and in seconds.  The IDE also provides clear outlines showing
implemented methods and signatures as collapsible outlines, eliminating
any problems with being able to quickly understand the structure and
details of a class.  My experience has been that any claimed time saved
by using Lombok, both in creating the generated code and in the
cognitive load of viewing and understanding data objects, is vastly
outweighed by the time lost debugging obscure build synchronization
issues from the willful misuse of annotation processors.