Complete Challenge 1.
This adds an address table and references it from the user table.

There are a couple changes in this that may be unexpected or raise
questions (i.e. controversial).  The biggest is that I rewrote the
UserService to use direct JDBC instead of Spring's JdbcTemplate.

When introducing a second table with a foreign key relationship, we move
from trivial mappings to non-trivial mappings of database tables to Java
objects.  JdbcTemplate is really designed for trivial single-table
queries.  For producing object graphs, a full-blown JPA solution is more
appropriate.  However, with the small nature of the challenge as well as
my own personal preference, I opted to simply write the JDBC code myself
rather than introduce the additional complexity of an ORM.  It would be
entirely reasonable to switch to JPA in the future, particularly if the
object graph gets more complex.  For now, this will suffice.

Another possibly unexpected change is that I replaced the mocking in
UserServiceTest with a full SpringBootTest.  The reason is simple:  the
UserService exists only to map database results into Java objects.  If
we mock out that step then we are really testing nothing at all, except
perhaps the "no result" being mapped correctly.

SQL connection problems being mapped to a 503 should be discussed within
the team.